telephony: trigger call to createAllApnList on EVENT_RECORDS_LOADED
BZ: 154925

The createAllApnList() function is triggered by the IMSI
ready event. To create the APN list, we need to have the
SPN and the GID1 data information. When the IMSI ready
event occured, the reading of those data is triggered but
not necessarily completed. Due to this the APNs of MVNOs
are not correctly retrieved.

This patch update the code to create the APN list on the
records loaded event.

Change-Id: I7a793891b9518d2ada9483698d9a7ae4755f0e80
Category: aosp improvement
Domain: Telephony-Telephony-FWK
Origin: internal
Upstream-Candidate: no, proprietary
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>