Fix issues seen in sim sms operations
BZ: 158326

Copy of messages to SIM SMS storage fails as call to
enforceCallingPermission results in SecurityException.
Exception is thrown as the copy to the SIM SMS storage
is called in the same process as phone.

Fix the issue by replacing the use of enforceCallingPermission
by enforceCallingOrSelfPermission.

Change-Id: I7d897b1fd9ffb6344be8daa6f67d3d29e2373205
Category: aosp improvement
Domain: Telephony-Telephony-FWK
Origin: internal
Upstream-Candidate: yes
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>