bip: add check of mToBeUsedApnType
BZ: 158967

In some case, mToBeUsedApnType can be null when we receive
the disconnect event. We need to check it before to use
it in the onDisconnected() function to avoid a JavaCrash
error.

This patch fixes the issue by adding the necessary check.

Change-Id: Ic8fc19cb29e84885d7c4b70bd1fc9da81e85951a
Category: aosp improvement
Domain: Telephony-Telephony-FWK
Origin: internal
Upstream-Candidate: no, proprietary
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>