[PORT FROM MAIN KK-MR1] Telephony: Fix for Failing OMA test SUPL-1.0-con-13
BZ: 55252

The WAP Push message from an SLP (SLC for non-proxy mode) to
a PPG shall contain the SUPL INIT message and shall follow
[WAP PAP]. When the device receive a NI request, device should
not answer to this request. Currently, this is answered resulting
in test failure.

Fix the issue by not forwarding the WAP Push SMS to APS if
the app id is wrong.

Change-Id: I4c3d5943fcf426a66e0a6003a7c6ce9f08691a28
Orig-Change-Id: Ifab87a40778f8ffea9f4991516352efe647fa80e
Category: aosp improvement
Domain: Telephony-Telephony-FWK
Origin: internal
Upstream-Candidate: yes
Signed-off-by: Jerome Pantaloni <jeromex.pantaloni@intel.com>
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>