Telephony: Fix for Failing OMA test SUPL-1.0-con-13
The WAP Push message from an SLP (SLC for non-proxy mode) to
a PPG shall contain the SUPL INIT message and shall follow
[WAP PAP]. When the device receive a NI request, device should
not answer to this request. Currently, this is answered resulting
in test failure.

Fix the issue by not forwarding the WAP Push SMS to APS if
the app id is wrong.

Change-Id: I4c3d5943fcf426a66e0a6003a7c6ce9f08691a28
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-9943
Category: aosp improvement
Domain: Telephony-Telephony-FWK
Origin: internal
Upstream-Candidate: yes
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: https://android.intel.com:443/299917
Reviewed-on: https://android.intel.com:443/303840