Fix issue in opening channel and sending channel status
BZ: 151494

Conformance test 31.124 27.22.4.31/1.3 fails as the
channel status envelope is not sent to SATK after link
is dropped.

If the link is dropped, requested APN type can still
be in RETRYING state. So, ignoring the disconnect event
when the APN is in retrying state will result in channel
status not sent to SATK and also BIP channels not cleared
properly.

Following changes are done to fix the issue:
- Fix the issue by removing the code which ignores the
disconnect event if mIsDataCallRetry is set or if the
disconnect is received when the requested APN is in
RETRYING state.
- If the Disconnect is received when the APN state
before enabling the APN is RETRYING or CONNECTING, then
enable the APN again.

Change-Id: I0a7cd8a8e7e39d2f1c5fa384f4ae6010f6bda38b
Category: device enablement
Domain: Telephony-Common
Origin: internal
Upstream-Candidate: no, proprietary
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>