Telephony: Don't use IA apn for Initial Attach if apn name is empty
BZ: 164539

When the apns-conf.xml has the apn settings where the IA apn
is set as an empty string (""), the LTE registration is not
working. With IA apn set to empty, the initialAttach uses an
empty apn string resulting in inconsistant data browsing and
also possibility of charging differently by some networks
because of a wrong APN usage.

This issue is fixed by ignoring the IA apn if the apn name is
empty and use the default apn instead.

Change-Id: Idfd3af5936271a52a9b7f143fe673f7cf59e84d5
Category: device enablement
Domain: Telephony-Common
Origin: internal
Upstream-Candidate: no, proprietary
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>