Fix issue in showing network information
BZ: 163050

When the device doesn't support voice but supports
data, sms and if registered to only GSM, signal
strength is not shown and also service state is shown
as No service.

Copy the Data registration state to voice registration
state if the device doesn't support voice or
voice registration state is not in service and data
registration state is in service. Also with this patch,
mVoiceCapable is used in the GsmServiceStateTracker instead
of reading the confiuration property value every time when
needed.

Change-Id: I23d52af6098a133548938af3a980406c5e357ba0
Category: aosp improvement
Domain: Telephony-Telephony-FWK
Origin: internal
Upstream-Candidate: yes
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>