telephony: fix clockwork issue.
BZ: 168655

ComprehensionTlv class has been enhanced to support tlv that is
container of childs tlvs. In this case, the rawvalue is not valid
until the tlv is encoded.
In order to prevent a direct null dereference, when null is passed
as rawValue in constructor, the member is assigned an empty byte
array.

Change-Id: I38bfc779b3b241687786af754aeea38f77cfc55c
Category: feature differentiation
Domain: Telephony-IMS-IR92
Origin: internal
Upstream-Candidate: no, proprietary
Signed-off-by: Alexandre d'Alton <alexandre.d'alton@intel.com>