stk: update response of open channel according to 3GPP spec
BZ: 174604

In order to make 31.124 27.22.4.27.6 SEQ 6.2 OPEN CHANNEL
LTE test case pass, the coding of Bearer description tag
should be updated to be compliant with the latest release
of the specification.

In case of a non-GBR QCI, the QoS octets in the "EPS
quality of service" infrmation are ignored by the UE, as
specified in the TS 24.301. This is not the case in the
current implmentation where those octets are always taken
into account.

This patch updates the code to ignore the QoS octets for
non-GBR QCI and be compliant with the latest release of
the 3GPP specifications.

Change-Id: I78de8c6ab206fdcf02746b7cff637d4ab23b80c1
Category: aosp improvement
Domain: Telephony-Telephony-FWK
Origin: internal
Upstream-Candidate: no, proprietary
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>