Fix list of apns disappearing on airplane mode deactivation
BZ: 174405

Upon deactivation of airplane mode with a pin enabled
sim card, list of apns under Access Point Names menu
disappears. List of apns reappear as soon as the
IMSI is read properly.

Upon SIM ready, SIM lock state is queried. If the lock is
active, then the pinLockedRegistrants are notified. This
results in SIM records resetting even when the pin1state
is unknown. As the IMSI is reset on resetting of SIM
records, list of APN disappears.

Fix the issue by resetting the SIM records only when the
previous app state is ready and the current app state is
puk locked.

Change-Id: I0369d70c94a2791aa418f29289e38dd607c16ca1
Category: aosp improvement
Domain: Telephony-Telephony-FWK
Origin: internal
Upstream-Candidate: yes
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>