Fix SPN name shown in lock screen even in out of service area
BZ: 167536

SPN name is shown in the lock screen when the device is
not registered to any network.

Display of SPN name is shown based on the information
in SIM. Since the information stored in the SIM card
is to show the SPN name, SPN name is shown in the lock
screen along with "NO SERVICE" text.

Override the SPN if
- Device is out of service
- Device is emergency registered but doesn't support voice.

Change-Id: Icb4625de8bed043683a2237f36159be85e105b09
Category: aosp improvement
Domain: Telephony-Telephony-FWK
Origin: internal
Upstream-Candidate: yes
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>