Fix for conformance test 31.124 27.22.4.27.6 SEQ 6.2 failure
BZ: 168363

Conformance test case 3GPP 31.124, 27.22.4.27.6 SEQ 6.2
failing as the PDN Connectivity request is issued
with wrong APN name.

Upon OPEN_CHANNEL proactive command, BIPGateway will
request for enabling the apn by calling enableApnType.
enableApnType will enable the requested APN with the
first APN stored in the waitingApnList. If there are
multiple ApnSetting which can handle the requested
ApnType, then data connection will be established with
the first apn parameters in the waiting apn list.
This results in PDN connectivity request established
with always the first apn in the list whereas the
conformance test is expecting the PDN connectivity request
with the requested apn.

Following changes are done to address this issue:
- Store the requested APN for BIP in the DcTrackerBase.
- Reset the stored BIP APN on connection failure,
CLOSE_CHANNEL.
- Add the ApnSetting to the waiting apn list only
if the requestedApnType is not APN_TYPE_CBS or
only if the requestedApn matches the apn of the ApnSetting.

Change-Id: Idaaa40489c2ab9993676282ffd51c429aaf0037e
Category: device enablement
Domain: Telephony-Common
Origin: internal
Upstream-Candidate: no, proprietary
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>