Clear roaming indicator when radio is off or unavailable
BZ: 188105

In some case, the roaming indication is still visible when
we do the following sequence:

 1. Switch air plane mode to ON (phone is in roaming)
 2. Remove the SIM card
 3. Switch air plane mode to OFF
 4. The roaming indiation is still visible

This issue is due to the fact, that the roaming indicator
is not cleared when the radio is switched off or not
available.

This patch fixes the issue by setting the roaming variables
to false when the radio is off or not available.

Change-Id: I301d8aac770b1b93d1a903b026356b12f6aed725
Category: aosp improvement
Domain: Telephony-Telephony-FWK
Origin: internal
Upstream-Candidate: yes
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>