Correct the issue while extracting mcc code
BZ: 198547

While converting the operator numeric to text string
the possibility of the operator numeric being empty thus
resulting in an exception is taken care of, but the member
variable mGotCountryCode is not correctly set. This could
result in wrong information being passed around in telephony
framework.

This patch fixes this issue by ensuring that mGotCountryCode
is set to true only after successfull conversion of the operator
numeric to text string.

Change-Id: Ie1cfcf0f8bcf3726830394441970ba1adb7fa453
Category: aosp improvement
Domain: Telephony-Telephony-FWK
Origin: internal
Upstream-Candidate: yes
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>