Telephony: Rework suspend/resume handling
BZ: 202021

Android framework decides concurrent voice and data
support based on the CS registration network type. In telephony
framework, if the network type is UMTS or above, it is
considered to not to support voice-data concurency. But there is
an issue with the network type values, where the value for GSM
will also be reported to support concurrency. See ril.h

For the use case mentioned in the BZ, the Ui display is wrong
when the data enable/disable setting is not matching with the
actual data state, during voice call.

The AOSP is modified to decide the voice-data concurency based
on the MS CLASS info from OemTelephony instead of the network
type values.

Change-Id: I91fdb9c72f17a44151a1d8a4efbf070c23e1e0b3
Category: aosp improvement
Domain: Telephony-Telephony-FWK
Origin: internal
Upstream-Candidate: no, proprietary
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>